Initial commit.
Simple and fairly naive implementation of component based architecture
which uses channels for communication. Current shortcomings include:
 * Component interaction is not yet entirely restricted to channels.
 * There are identification/uniqueness issues.

Overall, everything is very simple. A game Entity is created which
creates a player Entity and gives it a move Action. Over the next few
ticks, the Action executes and updates the player's Position state.

A good start though!