Better abstraction for Action.Act()
No longer passing Entity data structures for actions to modify.

CmpData now fulfills the Entity interface so that Action.Act() can
take an Entity arg. Possibly this will cause problems down the road as
the Id and Name are meaningless, in which case CmpData would need a
reference to the parent Entity, or Act() would take a CmpData as
argument. Either way seems kludgy.