Added PubSub service.
Not technically an envisioned service yet as it doesn't satisfy an
interface with that name (since there isn't one). This package allows
the creation of a publish/subscribe service that is data agnostic.

In addition, added a MsgId type to the msgId package and a Publish and
Subscribe message to the list of ids. By sharing this id space, channels
can be unified around one message type which should simplify
communications.