Changed Id() to Name() in ServiceMsg interface
Id has become no longer useful and it remains to be seen whether Name
will actually be useful, because possibly the string that Name returns
could be gotten by reflection. Also, if keeping Name *is* useful it
might make more sense to name it String to satisfy that interface.

For now at least Name allows ServiceMsg to be somewhat unique and this
may be useful in preventing errors, rather than just passing
interface{}. We shall see.