Fixed deadlock in gameloop, closes #9
observer was iterating through keys and not values in its list of entity
and view associations. Thus when getting a MsgTick, it passed the
message to entities rather than views and entity was getting a tick thus
replying to game early.

Also fixed sending on a closed channel on a view's second iteration. It
now makes a new channel every iteration. This was not seen before because
we had never gotten this far.