Implemented better error handling in comm
If a fatal error is encountered in a client loop, instead of panicking
we remove the client from the comm service and *then* exit the loop.
Because of this change, we no longer get a deadlock in the comm service
when dealing with 'broken' clients (who may have differing ideas about the
protocol).

connect() and tests still panic as before but use some utility functions
now, thus their error handling remains simple.