Shut down observer and fixed deadlock
Observers shut down now when their client disconnects. Before this
commit, a deadlock was caused by disconnected clients leaving views
trying to send to comm, but unable because nothing was draining the
channel. This caused entities unable to continue replying to state
updates and game was stuck trying to tick. This has been fixed by:

 * Shutting down observer from the comm service when a removeClient
   message is received.
 * Creating a goroutine which drains the client's (struct) SendQueue so
   that nothing else gets hung up.