Converted Game from entity to service
It has become apparent that it is more logical to treat and organize the
nexus of gameplay as a service rather than an entity. While the
flexibility of having it as an entity is interesting, it has always been
a special entity (different main loop and additional functions) and
working with its internal data has always been cumbersome. Further
support for this new treatment can be found in the fact that we pass
Game as part of the standard service context struct. All the other
channels in this context have real services listening on the other end
(with more planned). Thus it has already been acting somewhat as a
service and as mentioned before, very little like an entity.

Also removed old testing output from the main function.