Allow external access to entity input channel
The input channel for a given entity can now be accessed with Chan(),
this simplifies a good deal of code which was forced to treat entities
and their channels as separate. It also allows access to this channel
where it has not explicitly been passed (Action.Act was the motivator
here).