Generalized entity spawning in Game
Previously there specific functionality for spawning a player, this has
been generalized to spawn any entity which simplifies the code a bit and
allows us to spawn any entity through a message easily.

To make this possible, the entity creation functions have been changed to
'InitX' functions which return a core.Entity so the entity to spawn can
be specified simply by sending a function.

The modified/new MsgSpawnEntity lives in the game package as only game
handles it. It is not clear whether making messages package specific
like this clears their intention of use or makes them hard to find
rather than being with all the core messages.