Fixed MsgBuffer sending behavior and simplified func
The last implementation would only send a value when it received one
which was filling the buffer pretty quickly (depending on the volume of
messages received). This current implementation is simpler, but has the
drawback of always putting the received message in the queue.

The last implementation also had a bug where it would miss the
occasional message. This had the effect of preventing the player from
moving when other messages were being sent to World.