Added new concurrency utilities and converted most code to use them
The new concurrency handling operates off the simple principle that when
a given goroutine sends/receives to/from a channel, it must also
attempt to receive on its main input channel. This prevents a lot of
deadlock situations where an entity and service (for example) are
trying to send to eachother and thus neither send can go through.

The current behavior is to queue messages when blocked on
sending/receiving to/from a specific channel. These messages will then
be handled when the operation on the specific channel has gone through.
This behavior is easy to change for any given type implementing
MsgHandler.

Currently there is a possible deadlock when waiting for a reply to
return, as any messages that the waiting-on-reply goroutine may need to
handle before the whole game can proceed just sit in the queue.